refactor(config)!: implement a universal `metric_config` for all metrics
The `metric_config` class maps to the `tool_monitoring_config` table; instances map to rows in that table.
All DB manipulations related to metric configuration is encapsulated in that `metric_config` class.
This means neither the manager, nor any output classes need to concern themselves directly with the database.

Since an entry in the general config table is eventually required for _every_ registered metric, a `$config` property makes sense on the base `metric` class itself.
The `$config` property is guaranteed to be set before access (via the magic `__get` and later hopefully a proper hook).
Default implementations for the methods that define a metric-specific config form and its default values are now present on the `metric` base class, making the `configurable_metric` class obsolete.

Because the qualified metric name (component & name) is guaranteed by a DB constraint (and index) to be unique, the `configure` URL for a metric can use that qualified name as the parameter instead of the primary key/ID.
This greatly simplifies the way the manager has to store information about the registered metrics because the ID is no longer relevant anywhere (outside of the DB itself).
It is also nicer to use an identifier with semantic meaning.

The manager now also does not need to concern itself directly with the configs of individual metrics.
By default, any access to a metric via the manager has its config refreshed, i.e. fetched from the DB.