refactor!: merge manager into hook and merge config into metric
Since every metric ultimately needs one corresponding entry in the database table, the table is now simply called `tool_monitoring_metrics`.
The former `data` field in the table is now called `config`.

The public properties of a `metric` object now map exactly to the fields in the DB table and the `metric` class fully encapsultes all DB manipulations that directly relate to it.
For better visual separation from the monitoring logic, all DB manipulation methods reside in the `metric_orm` trait used by `metric`. Those methods are all private and thus inaccessible to subclasses.

There are two new public methods on `metric`:
- `save_config` can be called after the `config` or `enabled` properties have been changed from the outside.
- `register` (static) is the only public constructor for a metric that ensures a DB entry exists for the metric.

The latter is also the recommended callback to register a metric in the `db/hooks.php`.

The manager is now also the hook and is implemented as a singleton.
Metrics can only be added to it, but not removed.
Registered metrics can still be retrieved from it as before.

For static analysis tools and IDEs, the `metric` class is also generic in terms of the `config` type.
The `num_users_accessed` class makes use of this by declaring the `config` to have a `timewindow` property.