refactor(config)!: move all database responsibility out of `metric`
There is now a separate `registered_metric` class that takes on all responsibilities for metric-related DB manipulations.
A `registered_metric` acts as a wrapper for a `metric`.
No other parts of the code directly interact with `metric` instances.
The `metric` is no longer iterable, but the `registered_metric` is.
The `calculate` method of a `metric` is now public and the optional metric-specific config is now passed into that method as its only argument.
The `metric::get_config_form` method no longer exists because it would introduce a dependency on the `config` form class and the `add_config_form_elements` method should suffice in any case.
The `metric::validate_value` method is now public.
A nullary constructor for `metric` is now explicit, protected and final.

The `metrics_manager` is now distinct from the hook again. The hook is now called `metric_collection`.
The manager holds and hands out `registered_metric` objects.

The `save` method of the `config` form no longer requires the metric instance to be passed as an argument.

The base `metric` class and the hook are intentionally as minimal as possible with no dependencies on any other parts of the plugin.