refactor!: move fetching and synchronization to manager
The `metrics_manager` is no longer a singleton and accepts a `metric_collection` instance during construction.
That `metric_collection` hook is no longer dispatched in the manager's constructor but in a separate `dispatch_hook` method to make the intent clear.
The manager now has the `fetch` and `sync` methods instead of `get_enabled_metrics` and `sync_registered_metrics`.
Both `fetch` and `sync` first collect the metrics via the hook by default.
This can be disabled by passing `collect: false` to them, which (together with passing a custom `metric_collection` to the constructor) massively improves testability.
Corresponding methods are removed from the `registered_metric` class and unit tests are adjusted accordingly.

The `config` field is now nullable again for registered metrics and the `collect` method on the base `metric` reflects this by accepting `object|null`.